gateway/requirements.txt
fastapi==0.95.2
pydantic==1.10.9
confluent-kafka==1.9.2
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.22.0
pydantic[Email]==1.10.9